+++ src/main/ruby/truffleruby/core/array.rb
@@ -304,6 +304,7 @@ class Array
   def count(item = undefined)
     seq = 0
     if !Primitive.undefined?(item)
+      warn 'given block not used', uplevel: 1 if block_given?
       each { |o| seq += 1 if item == o }
     elsif block_given?
       each { |o| seq += 1 if yield(o) }
@@ -697,10 +698,8 @@ class Array
 
   def keep_if(&block)
     return to_enum(:keep_if) { size } unless block_given?
-
-    Primitive.check_frozen self
-
-    Primitive.steal_array_storage(self, select(&block))
+    select!(&block)
+    self
   end
 
   def last(n=undefined)
@@ -1205,8 +1204,24 @@ class Array
 
     Primitive.check_frozen self
 
-    ary = select(&block)
-    Primitive.steal_array_storage(self, ary) unless size == ary.size
+    original_size = size
+    selected_size = 0
+
+    each_with_index do |e, i|
+      abnormal_termination = true
+      to_select = yield e
+      abnormal_termination = false
+
+      if to_select
+        self[selected_size] = e
+        selected_size += 1
+      end
+    ensure
+      self[selected_size..-1] = self[i..-1] if abnormal_termination
+    end
+
+    self[selected_size..-1] = []
+    self unless selected_size == original_size
   end
   alias_method :filter!, :select!
 
@@ -1244,9 +1259,9 @@ class Array
   end
 
   def sort_by!(&block)
-    Primitive.check_frozen self
-
     return to_enum(:sort_by!) { size } unless block_given?
+
+    Primitive.check_frozen self
 
     Primitive.steal_array_storage(self, sort_by(&block))
   end